+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Changelog
 
-<!--next-version-placeholder-->
+## V0.3.0 (16/08/2024)
+- Sinks are now updated
+    - No "strat field" sink
+    - An O1D and photolysis field is added and used as sink
+    - Photolysis field generated using TUV model
+- Dzz diffusion revisited as too low in tropics and too high at poles
 
-## v0.1.0 (21/06/2023)
-
-- First release of `malta`!
+## V0.2.2 (18/03/2024)
+- Transport updated 
+- In line with publication in JAMES 
 
 ## V0.1.3 (15/03/2024)
 
 - Update of transport
 - Timestep is made an optional keyword
+
+## v0.1.0 (21/06/2023)
+
+- First release of `malta`!
+